move AddOrEdit.tsx in molecules and use layout.css now instant of in code style
