+++ src/components/controllPanel/presentation/fragments/edit/fragments/ControlPanelEditChain.tsx
@@ -15,8 +15,8 @@ import { DavitButton } from "../../../../../atomic/buttons/DavitButton";
 import { DavitDeleteButton } from "../../../../../atomic/buttons/DavitDeleteButton";
 import { ChainDecisionDropDownButton } from "../../../../../atomic/dropdowns/ChainDecisionDropDown";
 import { ChainLinkDropDownButton } from "../../../../../atomic/dropdowns/ChainLinkDropDown";
-import { AddOrEdit } from "../../../../../common/fragments/AddOrEdit";
 import { DavitLabelTextfield } from "../../../../../common/fragments/DavitLabelTextfield";
+import { AddOrEdit } from "../../../../../molecules/AddOrEdit";
 import { ControlPanel } from "../common/ControlPanel";
 import { OptionField } from "../common/OptionField";
